Dedicated file for litter logic
